backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utilities
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Email (optional - für Verifikation)
# sendgrid==6.11.0

# Payments (optional - für Stripe)
# stripe==7.11.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Code Quality
flake8==7.0.0
black==24.1.1
